GafferCortexBindings : Move to src/GafferCortexModule
There's nothing here that needs to be public, so it should all be private.
