+++ src/GafferCortexModule/ParameterHandlerBinding.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2013, Image Engine Design Inc. All rights reserved.
+//  Copyright (c) 2011, Image Engine Design Inc. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -34,14 +34,14 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#ifndef GAFFERCORTEXBINDINGS_EXECUTABLEOPHOLDERBINDING_H
-#define GAFFERCORTEXBINDINGS_EXECUTABLEOPHOLDERBINDING_H
+#ifndef GAFFERCORTEXMODULE_PARAMETERHANDLERBINDING_H
+#define GAFFERCORTEXMODULE_PARAMETERHANDLERBINDING_H
 
-namespace GafferCortexBindings
+namespace GafferCortexModule
 {
 
-void bindExecutableOpHolder();
+void bindParameterHandler();
 
-} // namespace GafferCortexBindings
+} // namespace GafferCortexModule
 
-#endif // GAFFERCORTEXBINDINGS_EXECUTABLEOPHOLDERBINDING_H
+#endif // GAFFERCORTEXMODULE_PARAMETERHANDLERBINDING_H
